refactor: re-scope modules from "ethevent" to "deposit" above DB.
